+++ flowman-core/src/test/scala/com/dimajix/flowman/documentation/RelationCollectorTest.scala
@@ -26,6 +26,7 @@ import com.dimajix.flowman.graph.Graph
 import com.dimajix.flowman.graph.Linker
 import com.dimajix.flowman.model.Mapping
 import com.dimajix.flowman.model.MappingIdentifier
+import com.dimajix.flowman.model.MappingOutputIdentifier
 import com.dimajix.flowman.model.Project
 import com.dimajix.flowman.model.Prototype
 import com.dimajix.flowman.model.Relation
@@ -67,15 +68,19 @@ class RelationCollectorTest extends AnyFlatSpec with Matchers with MockFactory {
         val execution = session.execution
 
         (mappingTemplate1.instantiate _).expects(context,None).returns(mapping1)
-        (mapping1.identifier _).expects().returns(MappingIdentifier("project/m1"))
-        (mapping1.context _).expects().returns(context)
-        (mapping1.outputs _).expects().returns(Set("main"))
+        (mapping1.identifier _).expects().atLeastOnce().returns(MappingIdentifier("project/m1"))
+        (mapping1.context _).expects().atLeastOnce().returns(context)
+        (mapping1.outputs _).expects().atLeastOnce().returns(Set("main"))
+        (mapping1.inputs _).expects().returns(Set(MappingOutputIdentifier("m2")))
+        (mapping1.describe _).expects(*,*).returns(Map("main" -> StructType(Seq.empty)))
         (mapping1.link _).expects(*).onCall((l:Linker) => Some(1).foreach(_ => l.input(MappingIdentifier("m2"), "main")))
 
         (mappingTemplate2.instantiate _).expects(context,None).returns(mapping2)
-        (mapping2.identifier _).expects().returns(MappingIdentifier("project/m2"))
-        (mapping2.context _).expects().returns(context)
-        (mapping2.outputs _).expects().returns(Set("main"))
+        (mapping2.identifier _).expects().atLeastOnce().returns(MappingIdentifier("project/m2"))
+        (mapping2.context _).expects().atLeastOnce().returns(context)
+        (mapping2.inputs _).expects().atLeastOnce().returns(Set())
+        (mapping2.outputs _).expects().atLeastOnce().returns(Set("main"))
+        (mapping2.describe _).expects(*,*).returns(Map("main" -> StructType(Seq.empty)))
         (mapping2.link _).expects(*).onCall((l:Linker) => Some(1).foreach(_ => l.read(RelationIdentifier("src"), Map("pcol"-> SingleValue("part1")))))
 
         (sourceRelationTemplate.instantiate _).expects(context,None).returns(sourceRelation)
@@ -83,12 +88,14 @@ class RelationCollectorTest extends AnyFlatSpec with Matchers with MockFactory {
         (sourceRelation.name _).expects().returns("src")
         (sourceRelation.context _).expects().returns(context)
         (sourceRelation.link _).expects(*).returns(Unit)
+        (sourceRelation.describe _).expects(*,Map("pcol"-> SingleValue("part1"))).returns(StructType(Seq()))
 
         (targetRelationTemplate.instantiate _).expects(context,None).returns(targetRelation)
         (targetRelation.identifier _).expects().atLeastOnce().returns(RelationIdentifier("project/tgt"))
         (targetRelation.name _).expects().returns("tgt")
         (targetRelation.context _).expects().returns(context)
         (targetRelation.link _).expects(*).returns(Unit)
+        (targetRelation.describe _).expects(*,Map("outcol"-> SingleValue("part1"))).returns(StructType(Seq()))
 
         (targetTemplate.instantiate _).expects(context,None).returns(target)
         (target.context _).expects().returns(context)
@@ -99,7 +106,7 @@ class RelationCollectorTest extends AnyFlatSpec with Matchers with MockFactory {
 
         val graph = Graph.ofProject(session, project, Phase.BUILD)
 
-        (mapping1.identifier _).expects().atLeastOnce().returns(MappingIdentifier("project/m1"))
+        //(mapping1.identifier _).expects().atLeastOnce().returns(MappingIdentifier("project/m1"))
         //(mapping2.identifier _).expects().atLeastOnce().returns(MappingIdentifier("project/m2"))
         (mapping1.requires _).expects().returns(Set())
         (mapping2.requires _).expects().returns(Set())
@@ -109,14 +116,12 @@ class RelationCollectorTest extends AnyFlatSpec with Matchers with MockFactory {
         (sourceRelation.provides _).expects().returns(Set())
         (sourceRelation.requires _).expects().returns(Set())
         (sourceRelation.schema _).expects().returns(None)
-        (sourceRelation.describe _).expects(*,Map("pcol"-> SingleValue("part1"))).returns(StructType(Seq()))
 
         (targetRelation.description _).expects().atLeastOnce().returns(Some("target relation"))
         (targetRelation.documentation _).expects().returns(None)
         (targetRelation.provides _).expects().returns(Set())
         (targetRelation.requires _).expects().returns(Set())
         (targetRelation.schema _).expects().returns(None)
-        (targetRelation.describe _).expects(*,Map("outcol"-> SingleValue("part1"))).returns(StructType(Seq()))
 
         val collector = new RelationCollector()
         val projectDoc = collector.collect(execution, graph, ProjectDoc(project.name))